remove props from Manager.jsx and replace props of child components with useContext
